requirements.txt
# --- FastAPI & Server ---
fastapi==0.111.0
uvicorn[standard]==0.30.1

# --- PQC & Crypto ---
# oqs-python (replaces your mock crypto)
oqs==0.10.2
# For QRNG
qiskit==1.1.0
qiskit-aer==0.14.1
quantcrypt==1.0.1  # The new PQC library
cryptography==42.0.8 # quantcrypt dependency

# --- Databases ---
# MongoDB Driver
pymongo[srv]==4.7.3
# PostgreSQL Async Driver
asyncpg==0.29.0
# SQLAlchemy (Async ORM)
sqlalchemy[asyncio]==2.0.31

# --- Auth & Security ---
# JWT Token Management
python-jose[cryptography]==3.3.0
# Password Hashing
passlib[bcrypt]

# --- AI & Analytics ---
pandas==2.2.2
scikit-learn==1.5.0

# --- NEW RAG LIBRARIES ---
langchain==0.2.5
langchain_community==0.2.5
langchain-mongodb==0.1.7     # For MongoDB Atlas Vector Search
langchain-google-genai==1.0.6  # For Gemini
sentence-transformers==2.7.0   # For multilingual embeddings
ollama==0.2.1                  # For Phi-3
googletrans-py==4.0.0-rc1      # For language detection
pypdf==4.2.0                   # To read legal PDFs

# --- Utilities ---
# Environment variable management
pydantic-settings==2.3.4

# --- Testing ---
pytest==8.4.2